Refactor core backend
