+++ core_backend/app/admin/routers.py
@@ -7,8 +7,8 @@ from sqlalchemy import text
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from ..configs.app_config import ALIGN_SCORE_API, ALIGN_SCORE_METHOD
-from ..db.engine import get_async_session
+from ..config import ALIGN_SCORE_API, ALIGN_SCORE_METHOD
+from ..database import get_async_session
 from ..utils import get_http_client
 
 router = APIRouter()
