+++ core_backend/app/auth/__init__.py
@@ -1,0 +1,1 @@
+from .routers import router
