+++ core_backend/app/auth/routers.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 from fastapi.security import OAuth2PasswordRequestForm
 
-from ..auth import authenticate_user, create_access_token
+from .dependencies import authenticate_user, create_access_token
 
 router = APIRouter()
