+++ core_backend/app/contents/__init__.py
@@ -1,0 +1,1 @@
+from .routers import router
