+++ core_backend/app/llm_call/parse_input.py
@@ -5,16 +5,8 @@ These are functions that can be used to parse the input questions.
 from functools import wraps
 from typing import Any, Callable, Tuple
 
-from ..configs.app_config import STANDARD_FAILURE_MESSAGE
-from ..configs.llm_prompts import (
-    PARAPHRASE_FAILED_MESSAGE,
-    PARAPHRASE_INPUT,
-    TRANSLATE_FAILED_MESSAGE,
-    TRANSLATE_INPUT,
-    IdentifiedLanguage,
-    SafetyClassification,
-)
-from ..schemas import (
+from ..question_answer.config import STANDARD_FAILURE_MESSAGE
+from ..question_answer.schemas import (
     ErrorType,
     ResultState,
     UserQueryRefined,
@@ -22,6 +14,14 @@ from ..schemas import (
     UserQueryResponseError,
 )
 from ..utils import setup_logger
+from .llm_prompts import (
+    PARAPHRASE_FAILED_MESSAGE,
+    PARAPHRASE_INPUT,
+    TRANSLATE_FAILED_MESSAGE,
+    TRANSLATE_INPUT,
+    IdentifiedLanguage,
+    SafetyClassification,
+)
 from .utils import _ask_llm_async
 
 logger = setup_logger("INPUT RAILS")
@@ -69,9 +69,9 @@ async def _classify_safety(
                 error_type=ErrorType.QUERY_UNSAFE,
             )
             error_response.debug_info.update(response.debug_info)
-            error_response.debug_info[
-                "safety_classification"
-            ] = safety_classification.value
+            error_response.debug_info["safety_classification"] = (
+                safety_classification.value
+            )
             error_response.debug_info["query_text"] = question.query_text
             logger.info(
                 (
