+++ core_backend/app/question_answer/__init__.py
@@ -1,0 +1,1 @@
+from .routers import router
