+++ core_backend/app/whatsapp_qa/config.py
@@ -1,0 +1,5 @@
+import os
+
+# WhatsApp variables
+WHATSAPP_VERIFY_TOKEN = os.environ.get("WHATSAPP_VERIFY_TOKEN", "no-token-set")
+WHATSAPP_TOKEN = os.environ.get("WHATSAPP_TOKEN", "no-token-set")
