+++ core_backend/app/whatsapp_qa/schemas.py
@@ -1,0 +1,25 @@
+from datetime import datetime
+
+from pydantic import BaseModel
+
+
+class WhatsAppIncoming(BaseModel):
+    """
+    Pydantic model for Incoming WhatsApp message
+    """
+
+    phone_number: str
+    message: str
+    phone_id: str
+    msg_id_from_whatsapp: str
+
+
+class WhatsAppResponse(BaseModel):
+    """
+    Pydantic model for WhatsApp Response messages
+    """
+
+    incoming_id: int
+    response_text: str
+    response_datetime_utc: datetime
+    response_status: int
