+++ core_backend/tests/rails/test_llm_response_in_context.py
@@ -3,13 +3,14 @@ These tests check LLM response content validation functions.
 LLM response content validation functions are rails that check if the responses
 are based on the given context or not.
 """
+
 from pathlib import Path
 from typing import List, Literal, Tuple
 
 import pytest
 import yaml
 
-from core_backend.app.configs.app_config import ALIGN_SCORE_API, ALIGN_SCORE_THRESHOLD
+from core_backend.app.config import ALIGN_SCORE_API, ALIGN_SCORE_THRESHOLD
 from core_backend.app.llm_call.check_output import (
     _get_alignScore_score,
     _get_llm_align_score,
