+++ core_backend/tests/rails/test_safety.py
@@ -2,9 +2,9 @@ from pathlib import Path
 
 import pytest
 
-from core_backend.app.configs.llm_prompts import SafetyClassification
+from core_backend.app.llm_call.llm_prompts import SafetyClassification
 from core_backend.app.llm_call.parse_input import _classify_safety
-from core_backend.app.schemas import (
+from core_backend.app.question_answer.schemas import (
     ErrorType,
     UserQueryRefined,
     UserQueryResponse,
